create core/CacheManager and core/ServiceWorker
