+++ ui/main.reel/main.js
@@ -1,7 +1,7 @@
 
 var Component = require("montage/ui/component").Component,
     sharedMoviesService = require("core/tmdb-service").shared,
-    cacheManager = require('./cache-manager.js')
+    CacheManager = require('core/cache-manager.js').CacheManager;
 
 //TODO use details in toggle buttons
 //TODO do not use matte toggle buttons
@@ -14,22 +14,28 @@ exports.Main = Component.specialize({
             this.canDrawGate.setField("moviesLoaded", false);
             this._initialDataLoad = this.moviesService.load();
 
-
-            cacheManager.eventCache.checking = function () {
-                console.log('CacheManager', 'checking');
-            }
-            cacheManager.eventCache.updateReady = function () {
-                console.log('CacheManager', 'updateReady');
-                if (window.confirm('A new version is available. Install now!')) {
-                    window.location.reload();
-                }
+            // Add events
+            CacheManager.events.error = function (error) {
+                console.error('MainUpdate', 'error', error);
             };
-
-            cacheManager.eventCache.noupdate = function () {
-                console.log('CacheManager', 'noupdate');
+            CacheManager.events.cached = function () {
+                console.log('MainUpdate', 'up-to-date (cached)');
             };
-            cacheManager.listenToUpdate();
-            cacheManager.checkForUpdate();
+            CacheManager.events.noUpdate = function () {
+                console.log('MainUpdate', 'up-to-date (noUpdate)');
+            };
+            CacheManager.events.checking = function () {
+                console.log('MainUpdate', 'checking');
+            };
+            CacheManager.events.updateReady = function () {
+                console.log('MainUpdate', 'updateReady');
+            };
+            CacheManager.events.noupdate = function () {
+                console.log('MainUpdate', 'noupdate');
+            };
+            CacheManager.confirmOnUpdateReady = true;
+            CacheManager.listenToUpdate();
+            CacheManager.checkForUpdate();
         }
     },
